requirements.txt
langgraph==1.0.10
langchain==1.2.10
langchain-openai==1.1.10
langchain-community==0.4.1
wikipedia==1.4.0
pydantic==2.12.5
python-dotenv==1.2.1
google-search-results==2.4.2
fastapi==0.134.0
uvicorn==0.41.0